+++ packages/triple-web-utils/package.json
@@ -38,6 +38,9 @@
       "eslint"
     ]
   },
+  "dependencies": {
+    "ua-parser-js": "^1.0.36"
+  },
   "devDependencies": {
     "@titicaca/triple-web": "workspace:*"
   },
