+++ packages/triple-web-utils/src/index.ts
@@ -2,3 +2,4 @@ export * from './check-client-app'
 export * from './parse-client-app-metadata'
 export * from './regex'
 export * from './user'
+export * from './user-agent'
