+++ packages/triple-web-utils/src/user-agent.ts
@@ -1,0 +1,11 @@
+import { IResult } from 'ua-parser-js'
+
+export function isMobile(userAgent: IResult) {
+  const { device } = userAgent
+
+  if (device.type === 'mobile' || device.type === 'tablet') {
+    return true
+  } else {
+    return false
+  }
+}
