--- packages/triple-web/src/utils/user-agent.ts
@@ -1,8 +1,0 @@
-import { UserAgentValue } from '../types'
-
-const MOBILE_REGEX =
-  /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i
-
-export function validateMobile(userAgent: UserAgentValue['ua']): boolean {
-  return MOBILE_REGEX.test(userAgent)
-}
